rebased on master

+++ src/JMS/Serializer/SerializationContext.php
@@ -18,7 +18,6 @@
 
 namespace JMS\Serializer;
 
-use JMS\Serializer\Exception\LogicException;
 use JMS\Serializer\Exception\RuntimeException;
 use Metadata\MetadataFactoryInterface;
 
@@ -48,12 +47,18 @@ class SerializationContext extends Context
 
     public function startVisiting($object)
     {
+        if (! is_object($object)) {
+            return;
+        }
         $this->visitingSet->attach($object);
         $this->visitingStack->push($object);
     }
 
     public function stopVisiting($object)
     {
+        if (! is_object($object)) {
+            return;
+        }
         $this->visitingSet->detach($object);
         $poppedObject = $this->visitingStack->pop();
 
@@ -64,8 +69,8 @@ class SerializationContext extends Context
 
     public function isVisiting($object)
     {
-        if ( ! is_object($object)) {
-            throw new LogicException('Expected object but got '.gettype($object).'. Do you have the wrong @Type mapping or could this be a Doctrine many-to-many relation?');
+        if (! is_object($object)) {
+            return false;
         }
 
         return $this->visitingSet->contains($object);
